seperated helper SingletonMeta class into another file
